chore: move the models from schedule_engine to a new app operations
