+++ backend/realtime_engine/tasks.py
@@ -2,7 +2,7 @@ from celery import shared_task
 from messages.publish import databus_event
 from typing import Any
 from datetime import date, datetime, timedelta
-from schedule_engine.models import Vehicle, Operator, Run
+from operations.models import Vehicle, Operator, Run
 from feed.models import Feed, Trip
 
 
@@ -32,7 +32,7 @@ def initialize_run(run_data: dict[str, Any]) -> tuple[bool, str | None]:
     except Exception as e:
         databus_event(
             "RUN_INITIALIZATION_ERROR",
-            {"error": str(e), **run_data}, # Error with payload
+            {"error": str(e), **run_data},  # Error with payload
         )
         return (False, None)
 
@@ -61,7 +61,7 @@ def start_run(run_data: dict[str, Any]) -> bool:
 
 @shared_task(queue="realtime_engine")
 def end_run(run_data: dict[str, Any]) -> bool:
-    #Transitions an IN_PROGRESS run to COMPLETED. Returns False if the run is missing, in a non-completable state, or an erorr occurs.
+    # Transitions an IN_PROGRESS run to COMPLETED. Returns False if the run is missing, in a non-completable state, or an erorr occurs.
     run_id = run_data.get("run_id")
     if run_id in (None, ""):
         return False
@@ -81,9 +81,10 @@ def end_run(run_data: dict[str, Any]) -> bool:
         )
         return False
 
+
 @shared_task(queue="realtime_engine")
 def validate_run(run_data: dict[str, Any]) -> bool:
-# Are all required fields present and non-empty?
+    # Are all required fields present and non-empty?
     required = [
         "vehicle_id",
         "operator_id",
@@ -93,39 +94,39 @@ def validate_run(run_data: dict[str, Any]) -> bool:
         "shape_id",
         "start_date",
         "start_time",
-        "schedule_relationship"
+        "schedule_relationship",
     ]
     if any(run_data.get(field) in (None, "") for field in required):
         return False
-#  JSON often sends numbers as strings ("0" instead of 0). The database stores it as an integer, so we convert. If the value can't be converted (e.g. "abc"), reject.
+    #  JSON often sends numbers as strings ("0" instead of 0). The database stores it as an integer, so we convert. If the value can't be converted (e.g. "abc"), reject.
     try:
         direction_id = int(run_data["direction_id"])
     except (TypeError, ValueError):
         return False
-# Is the date today? (quiza inncesario)
+    # Is the date today? (quiza inncesario)
     try:
         submitted_date = datetime.strptime(run_data["start_date"], "%Y-%m-%d").date()
     except (TypeError, ValueError):
         return False
     if submitted_date != date.today():
         return False
-# Does the vehicle exist?
+    # Does the vehicle exist?
     if not Vehicle.objects.filter(id=run_data["vehicle_id"]).exists():
         return False
-# Does the operator exist?
+    # Does the operator exist?
     if not Operator.objects.filter(id=run_data["operator_id"]).exists():
         return False
-# Is the bus in another active run?
+    # Is the bus in another active run?
     if Run.objects.filter(
         vehicle_id=run_data["vehicle_id"],
         run_status="IN_PROGRESS",
     ).exists():
         return False
-# Is there a current GTFS Feed
+    # Is there a current GTFS Feed
     feed = Feed.objects.filter(is_current=True).first()
     if feed is None:
         return False
-# Does the trip exist in the current GTFS feed?
+    # Does the trip exist in the current GTFS feed?
     trip_exists = Trip.objects.filter(
         feed=feed,
         trip_id=run_data["trip_id"],
